+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] / 2022-02-23
+### Bug Fixes
+- Null GetFileVersionInfo set Warning
+### Changed
+- Add GetAssemblyName
+- Add GetComments & GetFileDescription
+- Add ShowInformation On MainProject
+
 ## [1.1.1] / 2022-02-23
 ### New Features
 - Compile Multiple Examples with EndWith `Name`
@@ -140,6 +148,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.1.2]: ../../compare/1.1.1...1.1.2
 [1.1.1]: ../../compare/1.1.0...1.1.1
 [1.1.0]: ../../compare/1.0.2...1.1.0
 [1.0.2]: ../../compare/1.0.1...1.0.2